FIX use pivot for "geom only" object placement FIX use container as "root" in grid container states - allows moving whole instanced set to be moved at once

+++ src/physics/size.h
@@ -6,6 +6,5 @@ namespace physics
 {
     typedef uint32_t physics_id;
     typedef uint16_t height_map_id;
-    typedef uint32_t material_id;
-
+    typedef uint16_t material_id;
 }
